@@ -1,6 +1,8 @@
-FactoryGirl.define do
+# frozen_string_literal: true
+
+FactoryBot.define do
   factory :admin_user do
     sequence(:email) { |i| "test#{i}@test.com" }
-    password 'password'
+    password { 'password' }
   end
 end
