@@ -1,4 +1,6 @@
-FactoryGirl.define do
+# frozen_string_literal: true
+
+FactoryBot.define do
   factory :organization do
     sequence(:name) { |i| "Corporation X#{i}" }
   end
