@@ -1,4 +1,6 @@
-FactoryGirl.define do
+# frozen_string_literal: true
+
+FactoryBot.define do
   factory :photo do
     sequence(:url) { |i| "https://www.test-#{i}.com" }
   end
