@@ -1,6 +1,8 @@
-FactoryGirl.define do
+# frozen_string_literal: true
+
+FactoryBot.define do
   factory :tag do
     post
-    value 'Hot Dang'
+    value { 'Hot Dang' }
   end
 end
