@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Chronolog
-  VERSION = "0.2.3"
+  VERSION = '0.3.0'
 end
