@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'rails/generators'
 require 'rails/generators/active_record'
 
@@ -5,7 +7,7 @@ module Chronolog
   class InstallGenerator < Rails::Generators::Base
     include ActiveRecord::Generators::Migration
 
-    source_root File.expand_path('../templates', __FILE__)
+    source_root File.expand_path('templates', __dir__)
 
     def install_models
       copy_file 'models/changeset.rb', 'app/models/chronolog/changeset.rb'
