@@ -1,16 +1,18 @@
+# frozen_string_literal: true
+
 module ActiveAdmin
   module ChronologHelper
     def change_summary(changeset)
       target = if changeset.changeable.present?
-        link_to changeset.identifier, [:admin, changeset.changeable]
-      else
-        changeset.identifier
+                 link_to changeset.identifier, [:admin, changeset.changeable]
+               else
+                 changeset.identifier
       end
 
       admin_user = if changeset.admin_user.present?
-        link_to changeset.admin_user, [:admin, changeset.admin_user]
-      else
-        'Deleted User'
+                     link_to changeset.admin_user, [:admin, changeset.admin_user]
+                   else
+                     'Deleted User'
       end
 
       action = changeset.action.gsub(/e$/, '')
