@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 RSpec.describe Chronolog::Changeset, type: :model do
@@ -9,12 +11,12 @@ RSpec.describe Chronolog::Changeset, type: :model do
 
   it { should validate_inclusion_of(:action).in_array(Chronolog::Model.actions) }
 
-  describe ".recent" do
+  describe '.recent' do
     let!(:oldest) { create(:changeset, created_at: 1.week.ago) }
     let!(:older)  { create(:changeset, created_at: 3.days.ago) }
     let!(:newest) { create(:changeset, created_at: 2.days.ago) }
 
-    it "returns changesets sorted by the most recent" do
+    it 'returns changesets sorted by the most recent' do
       expect(Chronolog::Changeset.recent).to eq [newest, older, oldest]
     end
   end
